Only allow tails with a prefix slash
When looking through the local dependencies of a component we picked the
one that matches a prefix for our require statement, and wanted to
support a tail (that would fetch a specific file in the component). But
we didn't enforce the `/` to be part of the tail so for example if we
had two components: ``test`` and ``test_again`` and they were both local
dependencies ``test_again`` would match as ``test`` and tail ``_again``.